refactor(roadmap): extract STAGES and componentize DSALearningRoadmap (#2982)
